The great credits overhaul of 2013.
Credits are now implemented as simple concrete models that have back-reference
relationships to the items for which they credit people.  This does not reduce
any SQL queries in the general case as the credits are subquery loaded, but
paves the way for a MASSIVE reduction of database boilerplate.

In addition, some riders on this commit (sorry!) including some deletion of
__init__.py magic (will follow through on this) and more usage of the details
view_helper.