Continuing cleanups for credits and metadata.
In this commit, the 'rdbms' modules in common and people are deleted and their
remaining relevant assets redirected elsewhere.  As well as this, some fixes
for metadata (most notably seasons) have been made; the metadata migration is
mostly done but the current code could be cleaner and a lot more concise.