Add friendly database down errors.
This change makes the website send views which raise database OperationalErrors
during processing to a new 'database_oops' view, which has a corresponding
template that explains that the database is down.

The schedule service system automatically now suppresses such errors, and
instead handles such situations as if the URY services were down.  This
consequently disables most of the database overhead on static pages, allowing
them to be viewed during database downtime.

Currently this error view is NOT waived when in any form of debug mode; this
should probably be fixed at some point.

TODO: Possibly change the status code to 503 Service Unavailable.