Streamline ShowDB admission constraints.
The 'has_showdb_entry' field on ShowType has been retired and all
its usage has been transferred over to 'is_public', because the
only occurrence of a public not-in-show-database show is filler
and filler is now no longer part of the ShowDB anyway as of this
website revision.

The corresponding database table entry should be deleted once the
old website is retired and any MyURY usage corrected.

Template changes are required to make ShowDB links work correctly.