Load in only active attachables.
This is a mammoth commit that's required pulling up all sorts of
things, and will likely introduce more bugs than it's fixed.  A
quick summary:

- Timeslots no longer have real credits; they load theirs in from
  shows and filter them down to the credits that are valid for
  them.  show_season_timeslot_credit can now be deleted;

- Timeslot finishes are now hybrid properties; this will
  eventually mean that they will be used directly in queries
  instead of the clunkier (start + duration) they expand into;

- Transients have grown a massive amount of new hybrid methods and
  properties for testing whether they are active in given
  conditions.  One of these allows arbitrary objects to be
  compared, searching for URY date/time attribute conventions;

- All attachables are now joined to their attachees via a special
  primaryjoin that filters using the aforementioned object
  transient hybrid-method, to ensure that only the active ones
  are part of the relationship.

This should fix the (un-phabricated) bug of credits not being
filtered by active-ness whilst simultaneously ensuring that
credits and other attachables are easy to get out of the database
and efficient in the amount of queries they use.

Hopefully.