Fix T148.
This mega-commit mainly consolidates the query loading logic in a
way such that it can be modified easily later, but the main
outstanding benefit is that the new logic fixes (or should fix)
the problem of searches not autoloading credits.

The new credit code is written so that it can be slapped on any
model query regardles of it actually having credits - it'll right
itself if it can't find credits (easier to ask forgiveness than
permission) and simply not load the credits in.

Note the difference between LOADING credits (in a query) and
GETTING credits (from the model): for example, Timeslots need
credits *loading* (specifically, the season's show's credits) but
the act of getting them does not directly relate to this as it
involves pre-sifting through the show credits to find the ones
that were active in the timeslot.

No unit tests yet, but I have done a quick check through relevant
pages and it seems to work.